// src/shared/api/generated/model/transferRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */

export interface TransferRequest {
  /** 출금계좌 ID (내 계좌) */
  withdrawalAccountId?: number;
  /** 입금계좌번호 (하이픈 없이) */
  depositAccountNumber?: string;
  /**
     * 이체금액. 1원 이상의 정수
     * @minimum 1
     */
  amount?: number;
  /**
     * 내 통장에 남길 표시내용. 최대 10자
     * @maxLength 10
     */
  myPassbookMemo?: string;
  /**
     * 상대 통장에 남길 표시내용. 최대 10자
     * @maxLength 10
     */
  recipientPassbookMemo?: string;
}